requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1

diffusers==0.32.0
transformers==4.46.0
accelerate==1.2.0
safetensors==0.4.5
huggingface-hub==0.26.2
peft==0.13.2

moviepy
pillow
edge-tts
requests
mysqlclient

# Django ecosystem
djangorestframework
django-cors-headers
djangorestframework-simplejwt
django-filter

# Async task queue
celery
redis
